requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.2.6
packaging==25.0
python-dotenv==1.1.1
scikit-learn==1.7.1
scipy==1.15.3
threadpoolctl==3.6.0
Werkzeug==3.1.3
